backend/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
python-multipart
python-dotenv
azure-storage-blob